requirements.txt
PyPDF2==2.12.1
pdf2image==1.16.3
Pillow==9.2.0
pdfkit==1.0.0
